Create multiple files to find out asymptomatic complexity.
