+++ .gitignore
@@ -1,0 +1,3 @@
+data_large.txt
+.idea
+result.json
